refactor(discover): split discoverWeekly service into per-stage modules
Relocate the 3,449-line Discover Weekly service into
services/discoverWeekly/ via the campaign-standard inheritance chain,
composing the unchanged public class and singleton with an identical
export surface. Bodies moved text-identical; private became protected
only where the chain requires. The deprecated source-scraping Prisma
retry contract test is replaced with runtime retry behavior coverage
per the repo testing contract.

Adds a module-boundary behavioral suite.

Refs #485
